+++ src/DungeonBrothers/Assets/Project/CodeBase/Gameplay/Services/Random/RandomService.cs
@@ -4,7 +4,7 @@ namespace Project.CodeBase.Gameplay.Services.Random
 {
     public class RandomService : IRandomService
     {
-        public float DoRandomInRange(float min, float max) =>
+        public int DoRandomInRange(int min, int max) =>
             UnityEngine.Random.Range(min, max);
         
         public bool DoFiftyFifty() => UnityEngine.Random.value > 0.5f;
